+++ tests/Format/TimeFormatterTestCase.php
@@ -8,15 +8,6 @@
  * with this source code in the file LICENSE.
  */
 
-/**
- * This file is part of the "litgroup/datetime" package.
- *
- * (c) Roman Shamritskiy <roman@litgroup.ru>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
 declare(strict_types = 1);
 
 namespace Test\LitGroup\Time\Format;